refactor(models): move CostSource, Usage, and TokenRates models to pricing module
