+++ src/linux_mcp_server/models.py
@@ -2,10 +2,8 @@ import typing as t
 
 from datetime import datetime
 from pathlib import Path
-from typing import Literal
 
 from pydantic import BaseModel
-from pydantic import ConfigDict
 from pydantic import Field
 from pydantic import field_serializer
 from pydantic import model_validator
@@ -200,21 +198,3 @@ class LogEntries(BaseModel):
     @field_serializer("unit", "path")
     def serialize_empty_as_null(self, value: str | Path | None) -> str | None:
         return str(value) if value else None
-
-
-### Cost models ###
-CostSource = Literal["api", "config", "models_dev", "fallback", "local"]
-
-
-class Usage(BaseModel):
-    input_tokens: int = 0
-    output_tokens: int = 0
-    cost: float | None = None
-
-
-class TokenRates(BaseModel):
-    model_config = ConfigDict(frozen=True)
-
-    input_per_token: float
-    output_per_token: float
-    source: CostSource
